refactor: moving store types to the dto folder and create Point3 type
