+++ src/viewerapi/Geometry.ts
@@ -1,0 +1,1 @@
+export type Point3 = { x: number; y: number; z: number }
